+++ lib/util/exporter_config.rb
@@ -78,6 +78,65 @@ module Util
     end
 
 
+    # Optional. The array of event names for exclusion. Applied before
+    # `events_to_include` is consulted.
+    # Only use the exclude or include list; don't use both.
+    #
+    # @return [Array] Of event names.
+    def events_to_exclude
+      @config_info['Events to Exclude']
+    end
+
+
+    # Optional. The array of event names for inclusion. Applied after
+    # `events_to_exclude` is consulted. Only if `events_to_exclude`
+    # is an empty array is this list consulted.
+    # Only use the exclude or include list; don't use both.
+    #
+    # @return [Array] Of event names.
+    def events_to_include
+      @config_info['Events to Include']
+    end
+
+
+    # Whether or not to scale request expiry based on how long the date
+    # range is. Longer date spans will receive longer expiry durations.
+    #
+    # @return [Boolean] Whether to scale request expiry duration.
+    def scale_request_expiry
+      @config_info['Scale Request Expiry']
+    end
+
+
+    # When `scale_request_expiry` is set to true, we'll multiply the number
+    # of days in the request range by this number to arrive at the expiry duration.
+    # Note however, that we will apply a minimum duration when scaling is in effect, so that
+    # request for one or two days of data don't prematurely timeout.
+    #
+    # @return [Float] The number of seconds. Might be fractional.
+    def scale_request_seconds_per_day
+      @config_info['Scale Request Seconds Per Day']
+    end
+
+
+    # When `scale_request_expiry` is set to false, we'll set the window of
+    # time that any request is valid, to exactly this value.
+    #
+    # @return [Fixnum] The number of seconds.
+    def unscaled_request_expiry_seconds
+      @config_info['Unscaled Request Expiry Duration in Seconds']
+    end
+
+
+    # When `scale_request_expiry` is set to true, we'll cap the window of
+    # time that any request is valid, to this maximum.
+    #
+    # @return [Fixnum] The number of seconds.
+    def max_scaled_request_expiry_seconds
+      @config_info['Maximum Scaled Request Expiry Duration in Seconds']
+    end
+
+
     # Loads our `Exporter.config.json` file to set the {#output_directory} instance variable.
     # @return [void]
     def load_configuration