+++ Changelog.md
@@ -1,3 +1,17 @@
+## 0.6.0 (2016-02-26)
+
+This version of the tool enhances the configuration subsystem, adding configuration keys:
+
+  * Create Sub-folders by Date
+  * Use Event Dates in Filenames
+  * Events to Exclude
+  * Events to Include
+  * Scale Request Expiry
+  * Scale Request Seconds Per Day
+  * Unscaled Request Expiry Duration in Seconds
+  * Maximum Scaled Request Expiry Duration in Seconds
+
+
 ## 0.5.0 (2015-07-30)
 
 This is the initial version of the tool.
@@ -6,6 +20,6 @@ Workarounds:
 
   - Not Applicable
 
-Bugfixes:
+Bug Fixes:
 
   - None